+++ src/index.ts
@@ -12,8 +12,8 @@ import { DocumentRegistry } from '@jupyterlab/docregistry';
 import ExecuteTimeWidget, { PLUGIN_NAME } from './ExecuteTimeWidget';
 
 class ExecuteTimeWidgetExtension implements DocumentRegistry.WidgetExtension {
-  constructor(tracker: INotebookTracker, settingRegistry: ISettingRegistry) {
-    this._settingRegistry = settingRegistry;
+  constructor(tracker: INotebookTracker, settings: ISettingRegistry.ISettings) {
+    this._settings = settings;
     this._tracker = tracker;
   }
 
@@ -22,10 +22,10 @@ class ExecuteTimeWidgetExtension implements DocumentRegistry.WidgetExtension {
     panel: NotebookPanel,
     context: DocumentRegistry.IContext<INotebookModel>
   ) {
-    return new ExecuteTimeWidget(panel, this._tracker, this._settingRegistry);
+    return new ExecuteTimeWidget(panel, this._tracker, this._settings);
   }
 
-  private _settingRegistry: ISettingRegistry;
+  private _settings: ISettingRegistry.ISettings;
   private _tracker: INotebookTracker;
 }
 
@@ -36,14 +36,38 @@ const extension: JupyterFrontEndPlugin<void> = {
   id: PLUGIN_NAME,
   autoStart: true,
   requires: [INotebookTracker, ISettingRegistry],
-  activate: (
+  activate: async (
     app: JupyterFrontEnd,
     tracker: INotebookTracker,
     settingRegistry: ISettingRegistry
   ) => {
+    let settings: ISettingRegistry.ISettings;
+    try {
+      settings = await settingRegistry.load(`${PLUGIN_NAME}:settings`);
+    } catch (err: unknown) {
+      console.error(
+        `jupyterlab-execute-time: Could not load settings, so did not active ${PLUGIN_NAME}: ${err}`
+      );
+      return;
+    }
+
+    // If the plugin is enabled, force recording of timing
+    // We only do this once (not on every settings update) in case the user tries to turn it off
+    if (settings.get('enabled').composite) {
+      settingRegistry.load('@jupyterlab/notebook-extension:tracker').then(
+        (nbSettings: ISettingRegistry.ISettings) =>
+          nbSettings.set('recordTiming', true),
+        (err: Error) => {
+          console.error(
+            `jupyterlab-execute-time: Could not force metadata recording: ${err}`
+          );
+        }
+      );
+    }
+
     app.docRegistry.addWidgetExtension(
       'Notebook',
-      new ExecuteTimeWidgetExtension(tracker, settingRegistry)
+      new ExecuteTimeWidgetExtension(tracker, settings)
     );
 
     // eslint-disable-next-line no-console